Organize TUI generated and presentation code
