+++ clients/tui/src/ui/app.test.ts
@@ -1,8 +1,8 @@
 import {createTestRenderer} from '@opentui/core/testing';
 import {afterEach, describe, expect, it} from 'vitest';
+import type {SessionController} from '../session-controller.js';
+import {initialSessionState, type SessionState} from '../session-model.js';
 import {createOpenTuiApp, type OpenTuiApp, promptPreview, toolOutputPreview} from './app.js';
-import type {SessionController} from './session-controller.js';
-import {initialSessionState, type SessionState} from './session-model.js';
 
 const cleanup: Array<() => void> = [];
