+++ tests/test_tui.py
@@ -385,5 +385,5 @@ def test_run_context_records_invocation_boundary(tmp_path):
 
 
 def test_committed_protocol_schema_matches_python_contract():
-    schema_path = Path("clients/tui/src/protocol.schema.json")
+    schema_path = Path("clients/tui/src/generated/protocol.schema.json")
     assert json.loads(schema_path.read_text()) == ProtocolDocument.model_json_schema()
